Move functionality from system runner to window focus runner
